// src/shared/models/genre.model.ts
export enum RuGenres {
  notSelected = 'Не выбрано',
  action = 'Боевик',
  comedy = 'Комедия',
  fantasy = 'Фэнтези',
  horror = 'Ужасы',
}

export enum EnGenres {
  notSelected = 'notSelected',
  action = 'action',
  comedy = 'comedy',
  fantasy = 'fantasy',
  horror = 'horror',
}